+++ framework/views/modalAnimator.js
@@ -1,0 +1,53 @@
+/*
+Copyright 2013-2015 ASIAL CORPORATION
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+*/
+
+(function() {
+  'use strict;';
+
+  var module = angular.module('onsen');
+
+  module.factory('ModalAnimator', function() {
+    var ModalAnimator = Class.extend({
+
+      delay: 0,
+
+      /**
+       * @param {Object} options
+       * @param {String} options.timing
+       * @param {Number} options.duration
+       * @param {Number} options.delay
+       */
+      init: function(options) {
+        options = options || {};
+
+        this.timing = options.timing || this.timing;
+        this.duration = options.duration !== undefined ? options.duration : this.duration;
+        this.delay = options.delay !== undefined ? options.delay : this.delay;
+      },
+
+      show: function(modal, callback) {
+        callback();
+      },
+
+      hide: function(modal, callback) {
+        callback();
+      }
+    });
+
+    return ModalAnimator;
+  });
+})();